requirements.txt
fastapi
langchain
langgraph
requests
python-dotenv
sqlalchemy
langchain-google-genai
uvicorn
asyncpg
passlib
python-jose
argon2_cffi
python-multipart